+++ node2/node_test.py
@@ -47,7 +47,7 @@ class Node:
 
         while True:
             socketNodes.listen(1)
-            socketNodes.settimeout(1)
+            #socketNodes.settimeout(1)
             try :
                 conn, addr1 = socketNodes.accept()
                 #print(addr)
@@ -55,6 +55,7 @@ class Node:
                 if data:
                     decriptedData = ast.literal_eval(data.decode('utf-8'))
                     addr = decriptedData[0]
+                    print(decriptedData[1])
                     try:
                         """
                         We want to know what kind of message we received
@@ -152,7 +153,7 @@ class Node:
                         socketNodes.connect((neighbour, 5003))
                         socketNodes.send(self.message)
                         #print(b'sending'+self.message)
-                        self.contactedIP[neighbour] = 'waiting'
+                        #self.contactedIP[neighbour] = 'waiting'
                         break
                     except TimeoutError:
                         pass
@@ -200,7 +201,7 @@ class Node:
         """
         
         if self.blockchain.waiting_blocks == []:
-            #self.confirmed.clear()
+            self.confirmed.clear()
             self.neighboursOk.clear()
             self.confirmed.append(addr)
             self.blockchain.putting_block(receivedBlock)
@@ -239,11 +240,14 @@ class Node:
             else:
                 self.blockchain.putting_block(receivedBlock)
                 self.blockchain.waiting_blocks = [self.blockchain.compare_blocks()]
-                self.message = self.setMessage((self.ip_address,{'Block': self.blockchain.waiting_blocks[0]}))
-                nodesMessage = Thread(target = self.runNodesMessage) #Problem. We kill the last thread even if it didn't accomplished the task
-                nodesMessage.setDaemon(True)
-                nodesMessage.start()
-                nodesMessage.join()
+                if self.blockchain.waiting_blocks[0] == receivedBlock:
+                    self.confirmed.clear()
+                    self.confirmed.append(addr)
+                    self.message = self.setMessage((self.ip_address,{'Block': self.blockchain.waiting_blocks[0]}))
+                    nodesMessage = Thread(target = self.runNodesMessage) #Problem. We kill the last thread even if it didn't accomplished the task
+                    nodesMessage.setDaemon(True)
+                    nodesMessage.start()
+                    nodesMessage.join()
